+++ Sibusten.Philomena.Client.Examples/EnumerateSearchQuery.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using Serilog;
 using Sibusten.Philomena.Client.Extensions;
 using Sibusten.Philomena.Client.Fluent.Images;
 using Sibusten.Philomena.Client.Images;
@@ -19,7 +20,7 @@ namespace Sibusten.Philomena.Client.Examples
             PhilomenaClient client = new PhilomenaClient("https://derpibooru.org");
             PhilomenaImageSearchBuilder search = client.SearchImages("fluttershy").WithMaxImages(5);
 
-            Console.WriteLine("Download query simple");
+            Log.Information("Download query simple");
 
             // Using download all method
             await search
@@ -28,7 +29,7 @@ namespace Sibusten.Philomena.Client.Examples
                     .WithImageFileDownloader(image => $"ExampleDownloads/EnumerateSearchQuery/{image.Id}.{image.Format}")
                 .BeginDownload();
 
-            Console.WriteLine("Download query with delegates, skipping existing images");
+            Log.Information("Download query with delegates, skipping existing images");
 
             // Use Linq to skip images already downloaded
             // Note that filters in the search query itself are preferred since they will provide better performance. Don't use Linq filtering for conditions like image score.
@@ -39,7 +40,7 @@ namespace Sibusten.Philomena.Client.Examples
                     .WithImageFileDownloader(GetFileForImage)
                 .BeginDownload();
 
-            Console.WriteLine("Downloading images explicitly");
+            Log.Information("Downloading images explicitly");
 
             // Explicitly looping over each image and saving
             await foreach (IPhilomenaImage image in search.BeginSearch())
@@ -49,7 +50,7 @@ namespace Sibusten.Philomena.Client.Examples
                 await image.DownloadToFile(filename);
             }
 
-            Console.WriteLine("Downloading with multiple threads and progress updates");
+            Log.Information("Downloading with multiple threads and progress updates");
 
             // Downloading with multiple threads and progress updates
             // Also skips downloaded images like before
@@ -95,7 +96,7 @@ namespace Sibusten.Philomena.Client.Examples
         {
             // Simple progress updating to console
             double percentDownloaded = (double)imageSearchProgressInfo.Processed / (double)imageSearchProgressInfo.Total * 100.0;
-            Console.WriteLine($"{imageSearchProgressInfo.Processed}/{imageSearchProgressInfo.Total} ({percentDownloaded:00.0}%)");
+            Log.Information("Progress update: {ImagesProcessed}/{ImagesTotal} ({PercentDownloaded:#0.0}%)", imageSearchProgressInfo.Processed, imageSearchProgressInfo.Total, percentDownloaded);
         }
     }
 }