+++ Sibusten.Philomena.Client/Images/Downloaders/ConditionalImageDownloader.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+using Sibusten.Philomena.Client.Logging;
 
 namespace Sibusten.Philomena.Client.Images.Downloaders
 {
@@ -8,18 +10,26 @@ namespace Sibusten.Philomena.Client.Images.Downloaders
 
     public class ConditionalImageDownloader : PhilomenaImageDownloader
     {
+        private ILogger _logger;
+
         private readonly IPhilomenaDownloader<IPhilomenaImage> _downloader;
         private readonly ShouldDownloadImageDelegate _shouldDownloadImage;
 
         public ConditionalImageDownloader(ShouldDownloadImageDelegate shouldDownloadImage, IPhilomenaDownloader<IPhilomenaImage> downloader)
         {
+            _logger = Logger.Factory.CreateLogger(GetType());
+
             _shouldDownloadImage = shouldDownloadImage;
             _downloader = downloader;
         }
 
         public override async Task Download(IPhilomenaImage downloadItem, CancellationToken cancellationToken = default, IProgress<DownloadProgressInfo>? progress = null)
         {
-            if (_shouldDownloadImage(downloadItem))
+            bool shouldDownloadImage = _shouldDownloadImage(downloadItem);
+
+            _logger.LogDebug("Condition for downloading image {ImageId}: {ConditionResult}", downloadItem.Id, shouldDownloadImage);
+
+            if (shouldDownloadImage)
             {
                 await _downloader.Download(downloadItem, cancellationToken, progress);
             }