+++ Sibusten.Philomena.Client/Images/PagedPhilomenaImageSearch.cs
@@ -3,14 +3,18 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using System.Threading;
+using Microsoft.Extensions.Logging;
 using Sibusten.Philomena.Api;
 using Sibusten.Philomena.Api.Models;
+using Sibusten.Philomena.Client.Logging;
 using Sibusten.Philomena.Client.Options;
 
 namespace Sibusten.Philomena.Client.Images
 {
     public class PagedPhilomenaImageSearch : IPhilomenaImageSearch
     {
+        private ILogger _logger;
+
         private const int _perPage = 50;
 
         private PhilomenaApi _api;
@@ -19,6 +23,8 @@ namespace Sibusten.Philomena.Client.Images
 
         public PagedPhilomenaImageSearch(PhilomenaApi api, string query, ImageSearchOptions? options)
         {
+            _logger = Logger.Factory.CreateLogger(GetType());
+
             _api = api;
             _query = query;
             _options = options ?? new ImageSearchOptions();
@@ -40,12 +46,18 @@ namespace Sibusten.Philomena.Client.Images
             if (_options.SortOptions?.SortField == SortField.Random && _options.SortOptions?.RandomSeed is null)
             {
                 _randomSeed = _api.GetRandomSearchSeed();  // TODO: Expose this method in the API interface
+
+                _logger.LogDebug("Search sort field is Random and a seed was not provided. Generated a new seed: {RandomSeed}", _randomSeed);
             }
+
+            _logger.LogDebug("Beginning image search for '{Query}'", _query);
 
             // Enumerate images
             ImageSearchModel search;
             do
             {
+                _logger.LogDebug("Downloading page {Page} of image search '{Query}'", page, _query);
+
                 // Get the current page of images
                 search = await _api.SearchImagesAsync(_query, page, _perPage, _options.SortOptions?.SortField, _options.SortOptions?.SortDirection, _options.FilterId, _options.ApiKey, _randomSeed, cancellationToken);
 
@@ -68,6 +80,8 @@ namespace Sibusten.Philomena.Client.Images
                 // Avoid reporting more metadata downloaded than total images to download
                 metadataDownloaded = Math.Min(metadataDownloaded, totalImagesToDownload);
 
+                _logger.LogDebug("Downloaded metadata for {MetadataDownloaded}/{TotalImagesToDownload} images", metadataDownloaded, totalImagesToDownload);
+
                 // Update metadata progress
                 metadataProgress?.Report(new()
                 {
@@ -82,14 +96,20 @@ namespace Sibusten.Philomena.Client.Images
 
                     if (image.IsSvgImage)
                     {
+                        _logger.LogDebug("Image {ImageId} is an SVG image", image.Id);
+
                         if (_options.SvgMode is SvgMode.RasterOnly or SvgMode.Both)
                         {
+                            _logger.LogDebug("Processing raster version of image {ImageId}", image.Id);
+
                             // Provide the original image which represents the raster version
                             yield return image;
                         }
 
                         if (_options.SvgMode is SvgMode.SvgOnly or SvgMode.Both)
                         {
+                            _logger.LogDebug("Processing SVG version of image {ImageId}", image.Id);
+
                             // Provide a new image which represents the SVG version
                             IPhilomenaImage svgImage = new PhilomenaImage(imageModel)
                             {
@@ -100,10 +120,14 @@ namespace Sibusten.Philomena.Client.Images
                     }
                     else
                     {
+                        _logger.LogDebug("Processing image {ImageId}", image.Id);
+
                         yield return image;
                     }
 
                     imagesProcessed++;
+
+                    _logger.LogDebug("Processed {ImagesProcessed}/{TotalImagesToProcess} images", imagesProcessed, totalImagesToDownload);
 
                     // Update search progress
                     searchProgress?.Report(new()
@@ -114,6 +138,8 @@ namespace Sibusten.Philomena.Client.Images
 
                     if (imagesProcessed >= _options.MaxImages)
                     {
+                        _logger.LogDebug("Image search '{Query}' stopping due to reaching the max number of images ({MaxImages})", _query, _options.MaxImages);
+
                         yield break;
                     }
                 }