+++ Sibusten.Philomena.Client/PhilomenaClient.cs
@@ -1,20 +1,26 @@
 ﻿using System;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
 using Sibusten.Philomena.Api;
 using Sibusten.Philomena.Api.Models;
 using Sibusten.Philomena.Client.Fluent.Images;
+using Sibusten.Philomena.Client.Logging;
 
 namespace Sibusten.Philomena.Client
 {
     public class PhilomenaClient : IPhilomenaClient
     {
+        private ILogger _logger;
+
         private PhilomenaApi _api;
 
         public string? ApiKey { get; set; } = null;
 
         public PhilomenaClient(string baseUrl)
         {
+            _logger = Logger.Factory.CreateLogger(GetType());
+
             _api = new PhilomenaApi(baseUrl);
         }
 
@@ -27,7 +33,9 @@ namespace Sibusten.Philomena.Client
         {
             string tagQuery = $"id:{tagId}";
 
-            TagSearchModel tagSearch = await _api.SearchTagsAsync(tagQuery, 1, 1);
+            _logger.LogDebug("Searching for tags: '{Query}'", tagQuery);
+
+            TagSearchModel tagSearch = await _api.SearchTagsAsync(tagQuery, page: 1, perPage: 1);
 
             if (tagSearch.Tags is null)
             {